refactor: reorganize test suite boundaries
