+++ tests/e2e_pytest/__init__.py
@@ -1,0 +1,1 @@
+"""Legacy pytest-based E2E scenarios that are not yet covered by Hurl."""
