+++ tests/e2e_pytest/test_available_models.py
