+++ tests/integration/__init__.py
@@ -1,0 +1,1 @@
+"""Integration and contract tests kept outside the black-box HTTP E2E suite."""
