+++ tests/integration/cache/__init__.py
@@ -1,0 +1,1 @@
+"""Cache integration tests."""
