+++ tests/integration/graphstorage/__init__.py
@@ -1,0 +1,1 @@
+"""Graph storage contract tests."""
