+++ tests/integration/graphstorage/test_neo4j_storage.py
@@ -11,9 +11,9 @@ import pytest
 import pytest_asyncio
 
 from aperag.graph.lightrag.kg.neo4j_sync_impl import Neo4JSyncStorage
-from tests.e2e_test.graphstorage.graph_storage_oracle import GraphStorageOracle
-from tests.e2e_test.graphstorage.networkx_baseline_storage import NetworkXBaselineStorage
-from tests.e2e_test.graphstorage.test_graph_storage import GraphStorageTestSuite, load_graph_data
+from tests.integration.graphstorage.graph_storage_oracle import GraphStorageOracle
+from tests.integration.graphstorage.networkx_baseline_storage import NetworkXBaselineStorage
+from tests.integration.graphstorage.test_graph_storage import GraphStorageTestSuite, load_graph_data
 
 dotenv.load_dotenv(".env")
