+++ tests/integration/graphstorage/test_postgres_graph_storage.py
@@ -11,9 +11,9 @@ import pytest
 import pytest_asyncio
 
 from aperag.graph.lightrag.kg.pg_ops_sync_graph_storage import PGOpsSyncGraphStorage
-from tests.e2e_test.graphstorage.graph_storage_oracle import GraphStorageOracle
-from tests.e2e_test.graphstorage.networkx_baseline_storage import NetworkXBaselineStorage
-from tests.e2e_test.graphstorage.test_graph_storage import GraphStorageTestSuite, load_graph_data
+from tests.integration.graphstorage.graph_storage_oracle import GraphStorageOracle
+from tests.integration.graphstorage.networkx_baseline_storage import NetworkXBaselineStorage
+from tests.integration.graphstorage.test_graph_storage import GraphStorageTestSuite, load_graph_data
 
 dotenv.load_dotenv(".env")
